+++ CHANGES.rst
@@ -1,12 +1,18 @@
-Changes in Matrix iOS SDK in 0.12.x (2019-xx-xx)
+Changes in Matrix iOS SDK in 0.12.6 (2019-05-)
 ===============================================
 
 Improvements:
- * MXRestClient/MXHTTPClient: support multiple SSL pinning modes (none/public key/certificate).
+ * MXHTTPClient: support multiple SSL pinning modes (none/public key/certificate)
  * MXHTTPClient: Enable the certificate pinning mode by default as soon as some certificates are present in the application bundle.
+ * MXEvent: Create a MXEventUnsignedData model for `MXEvent.unsignedData`.
+ * MXEvent: Add relatesTo property.
+ * Aggregations: Create MXSession.MXAggregations to manage Matrix aggregations API.
 
 Bug Fix:
  * MXMediaLoader: Disable trusting the built-in anchors certificates when the certificate pinning is enabled.
+
+API break:
+ * MXEvent: unsignedData is now of type MXEventUnsignedData.
 
 Changes in Matrix iOS SDK in 0.12.5 (2019-05-03)
 ===============================================